Added security controls on management services

+++ mystore-core/src/main/java/net/andresbustamante/mystore/core/services/CustomersManagementServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.extern.slf4j.Slf4j;
+import net.andresbustamante.mystore.api.exceptions.AccessDeniedException;
 import net.andresbustamante.mystore.api.exceptions.ApplicationException;
 import net.andresbustamante.mystore.api.exceptions.InvalidEmailException;
 import net.andresbustamante.mystore.api.exceptions.ObjectNotFoundException;
@@ -18,6 +19,7 @@ import net.andresbustamante.mystore.api.model.UserCreation;
 import net.andresbustamante.mystore.api.services.AddressesManagementService;
 import net.andresbustamante.mystore.api.services.CustomersManagementService;
 import net.andresbustamante.mystore.api.services.UsersManagementService;
+import net.andresbustamante.mystore.api.util.UserContext;
 import net.andresbustamante.mystore.jpa.dao.AddressDao;
 import net.andresbustamante.mystore.jpa.dao.CustomerDao;
 import net.andresbustamante.mystore.jpa.dao.UserDao;
@@ -47,7 +49,7 @@ public class CustomersManagementServiceImpl implements CustomersManagementServic
 
     @Override
     @Transactional(rollbackFor = ApplicationException.class)
-    public int createCustomer(@NonNull final CustomerCreation newCustomer) throws ApplicationException {
+    public int createCustomer(@NonNull final CustomerCreation newCustomer, final UserContext ctx) throws ApplicationException {
         if (customerDao.existsByEmail(newCustomer.email().toLowerCase(Locale.getDefault()))) {
             throw new InvalidEmailException("A customer already exists for the given email address");
         }
@@ -56,8 +58,8 @@ public class CustomersManagementServiceImpl implements CustomersManagementServic
         customer.setFirstName(newCustomer.firstName());
         customer.setLastName(newCustomer.lastName());
         customer.setEmail(newCustomer.email().toLowerCase(Locale.getDefault()));
-        customer.setAddress(createAddressForCustomer(newCustomer));
-        customer.setUser(createUserForCustomer(newCustomer));
+        customer.setAddress(createAddressForCustomer(newCustomer, ctx));
+        customer.setUser(createUserForCustomer(newCustomer, ctx));
 
         customer = customerDao.save(customer);
 
@@ -69,12 +71,17 @@ public class CustomersManagementServiceImpl implements CustomersManagementServic
 
     @Override
     @Transactional(rollbackFor = ApplicationException.class)
-    public void deactivateCustomer(@NonNull final Integer customerId) throws ApplicationException {
+    public void deactivateCustomer(@NonNull final Integer customerId, final UserContext ctx) throws ApplicationException {
         CustomerEntity customer = customerDao.findById(customerId).orElseThrow(
                 () -> new ObjectNotFoundException(String.format("Customer not found with the given ID: %d", customerId)));
 
+        if (!ctx.getUsername().equals(customer.getUser().getUsername())) {
+            // Only the same user can update himself
+            throw new AccessDeniedException("You cannot update this user");
+        }
+
         // Also deactivate the user to avoid any authentication attempt
-        usersManagementService.deactivateUser(customer.getUser().getId());
+        usersManagementService.deactivateUser(customer.getUser().getId(), ctx);
         customerDao.delete(customer); // Hibernate makes a soft-delete here
 
         log.info("Customer {} has been deactivated", customerId);
@@ -82,8 +89,14 @@ public class CustomersManagementServiceImpl implements CustomersManagementServic
 
     @Override
     @Transactional(rollbackFor = ApplicationException.class)
-    public void updateCustomer(final Integer customerId, final CustomerUpdate newCustomer) throws ApplicationException {
+    public void updateCustomer(final Integer customerId, final CustomerUpdate newCustomer, final UserContext ctx)
+            throws ApplicationException {
         CustomerEntity customer = customerDao.findById(customerId).orElseThrow();
+
+        if (!ctx.getUsername().equals(customer.getUser().getUsername())) {
+            // Only the same user can update himself
+            throw new AccessDeniedException("You cannot update this user");
+        }
 
         customer.setFirstName(newCustomer.firstName());
         customer.setLastName(newCustomer.lastName());
@@ -95,18 +108,20 @@ public class CustomersManagementServiceImpl implements CustomersManagementServic
         log.info("The customer {} has been successfully updated", customerId);
     }
 
-    private UserEntity createUserForCustomer(final CustomerCreation newCustomer) throws ApplicationException {
+    private UserEntity createUserForCustomer(final CustomerCreation newCustomer, final UserContext ctx)
+            throws ApplicationException {
         int userId = usersManagementService.createUser(new UserCreation(newCustomer.username(),
-                new String(newCustomer.password(),StandardCharsets.UTF_8)));
+                new String(newCustomer.password(),StandardCharsets.UTF_8)), ctx);
         return userDao.getReferenceById(userId);
     }
 
-    private AddressEntity createAddressForCustomer(final CustomerCreation newCustomer) throws ApplicationException {
+    private AddressEntity createAddressForCustomer(final CustomerCreation newCustomer, final UserContext ctx)
+            throws ApplicationException {
         AddressCreation newAddress = new AddressCreation(
                 newCustomer.addressLine1(), newCustomer.addressLine2(),
                 newCustomer.postalCode(), newCustomer.cityId());
 
-        int addressId = addressesManagementService.createAddress(newAddress);
+        int addressId = addressesManagementService.createAddress(newAddress, ctx);
 
         return addressDao.getReferenceById(addressId);
     }
